wip: split evm
