+++ src/db/mod.rs
@@ -4,7 +4,7 @@ pub mod sync;
 
 /// Database abstraction traits.
 mod traits;
-pub use traits::{ArcUpgradeError, CachingDb, StateAcc, TryCachingDb, TryStateAcc};
+pub use traits::{ArcUpgradeError, CachingDb, DbConnect, StateAcc, TryCachingDb, TryStateAcc};
 
 /// Cache-on-write database. A memory cache that caches only on write, not on
 /// read. Intended to wrap some other caching database.
