+++ src/evm/factory.rs
@@ -1,47 +1,12 @@
 use crate::{
-    helpers::Ctx, Block, Cfg, EvmErrored, EvmNeedsBlock, EvmNeedsCfg, EvmNeedsTx, EvmReady,
-    EvmTransacted, Tx,
+    db::DbConnect, helpers::Ctx, Block, Cfg, EvmErrored, EvmNeedsBlock, EvmNeedsCfg, EvmNeedsTx,
+    EvmReady, EvmTransacted, Tx,
 };
-use core::convert::Infallible;
 use revm::{
     context::result::{EVMError, ResultAndState},
     Database, Inspector,
 };
 use std::format;
-
-/// Trait for types that can be used to connect to a database.
-///
-/// Connectors should contain configuration information like filesystem paths.
-/// They are intended to enable parallel instantiation of multiple EVMs in
-/// multiple threads sharing some database configuration
-///
-/// The lifetime on this trait allows the resulting DB to borrow from the
-/// connector. E.g. the connector may contain some `Db` and the resulting Db may
-/// contain `&Db`. This allows for (e.g.) shared caches between DBs on multiple
-/// threads.
-pub trait DbConnect: Sync {
-    /// The database type returned when connecting.
-    type Database: Database;
-
-    /// The error type returned when connecting to the database.
-    type Error: core::error::Error;
-
-    /// Connect to the database.
-    fn connect(&self) -> Result<Self::Database, Self::Error>;
-}
-
-impl<Db> DbConnect for Db
-where
-    Db: Database + Clone + Sync,
-{
-    type Database = Self;
-
-    type Error = Infallible;
-
-    fn connect(&self) -> Result<Self::Database, Self::Error> {
-        Ok(self.clone())
-    }
-}
 
 /// Trait for types that can create EVM instances.
 ///
@@ -150,3 +115,53 @@ pub trait EvmFactory: DbConnect {
         }
     }
 }
+
+// Some code above and documentation is adapted from the revm crate, and is
+// reproduced here under the terms of the MIT license.
+//
+// MIT License
+//
+// Copyright (c) 2021-2024 draganrakita
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+
+// Some code above is reproduced from `reth`. It is reused here under the MIT
+// license.
+//
+// The MIT License (MIT)
+//
+// Copyright (c) 2022-2024 Reth Contributors
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
